+++ src/utils/Timeout.ts
@@ -4,12 +4,12 @@ export class Timeout {
     private _timeoutId: number | null = null;
 
     destroy(): void {
-        this._clearTimeout();
+        this.clearTimeout();
     }
 
     tick() {
         return new Promise<void>((resolve) => {
-            this._clearTimeout();
+            this.clearTimeout();
             this._timeoutId = GLib.timeout_add_seconds(GLib.PRIORITY_DEFAULT, 0, () => {
                 this._timeoutId = null;
                 resolve();
@@ -18,7 +18,18 @@ export class Timeout {
         });
     }
 
-    private _clearTimeout() {
+    once(milliseconds: number) {
+        return new Promise<void>((resolve) => {
+            this.clearTimeout();
+            this._timeoutId = GLib.timeout_add(GLib.PRIORITY_DEFAULT, milliseconds, () => {
+                this._timeoutId = null;
+                resolve();
+                return GLib.SOURCE_REMOVE;
+            });
+        });
+    }
+
+    clearTimeout() {
         if (this._timeoutId) {
             GLib.Source.remove(this._timeoutId);
             this._timeoutId = null;